requirements.txt
bcrypt==3.2.0
cryptography==3.2.1
flask-cors==3.0.9
flask-migrate==2.5.3
flask-restful==0.3.8
flask-script==2.0.6
pip-chill==1.0.0
psycopg2-binary==2.8.6
pyjwt==1.7.1
requests==2.24.0
